pyproject.toml
[project]
name = "news-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
